+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Constants/ModuleSettingOptionValues.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Constants;
+
+class ModuleSettingOptionValues
+{
+    /**
+     * Setting option values
+     */
+    public const NO_VALUE_ID = 0;
+}