+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/SchemaConfiguratorExecuters/SingleEndpointSchemaConfiguratorExecuter.php
@@ -1,0 +1,64 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Services\SchemaConfiguratorExecuters;
+
+use GraphQLAPI\GraphQLAPI\Constants\ModuleSettingOptions;
+use GraphQLAPI\GraphQLAPI\Constants\ModuleSettingOptionValues;
+use GraphQLAPI\GraphQLAPI\Facades\UserSettingsManagerFacade;
+use GraphQLAPI\GraphQLAPI\ModuleResolvers\SchemaConfigurationFunctionalityModuleResolver;
+use GraphQLAPI\GraphQLAPI\Registries\ModuleRegistryInterface;
+use GraphQLAPI\GraphQLAPI\Services\SchemaConfigurators\SchemaConfiguratorInterface;
+use GraphQLAPI\GraphQLAPI\Services\SchemaConfigurators\SingleEndpointSchemaConfigurator;
+use GraphQLByPoP\GraphQLEndpointForWP\EndpointHandlers\GraphQLEndpointHandler;
+use PoP\ComponentModel\Instances\InstanceManagerInterface;
+
+class SingleEndpointSchemaConfiguratorExecuter extends AbstractSchemaConfiguratorExecuter
+{
+    public function __construct(
+        InstanceManagerInterface $instanceManager,
+        protected ModuleRegistryInterface $moduleRegistry,
+        protected SingleEndpointSchemaConfigurator $endpointSchemaConfigurator,
+    ) {
+        parent::__construct(
+            $instanceManager,
+        );
+    }
+
+    /**
+     * This is the Schema Configuration ID
+     */
+    protected function getCustomPostID(): ?int
+    {
+        // Only enable it when executing a query against the single endpoint
+        /** @var GraphQLEndpointHandler */
+        $graphQLEndpointHandler = $this->instanceManager->getInstance(GraphQLEndpointHandler::class);
+        if (!$graphQLEndpointHandler->isEndpointRequested()) {
+            return null;
+        }
+        return $this->getUserSettingSchemaConfigurationID();
+    }
+
+    /**
+     * Return the stored Schema Configuration ID
+     */
+    protected function getUserSettingSchemaConfigurationID(): ?int
+    {
+        $userSettingsManager = UserSettingsManagerFacade::getInstance();
+        $schemaConfigurationID = $userSettingsManager->getSetting(
+            SchemaConfigurationFunctionalityModuleResolver::SCHEMA_CONFIGURATION,
+            ModuleSettingOptions::VALUE_FOR_SINGLE_ENDPOINT
+        );
+        // `null` is stored as OPTION_VALUE_NO_VALUE_ID
+        if ($schemaConfigurationID == ModuleSettingOptionValues::NO_VALUE_ID) {
+            return null;
+        }
+        return $schemaConfigurationID;
+    }
+
+    protected function getSchemaConfigurator(): SchemaConfiguratorInterface
+    {
+        return $this->endpointSchemaConfigurator;
+    }
+}